requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-apscheduler==0.6.0
mysqlclient==2.2.0
djangorestframework-api-key==2.1.0
apscheduler==3.10.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
